+++ weldx/__init__.py
@@ -49,6 +49,7 @@ These classes and functions are used to define welding processes.
 
     Time
     TimeSeries
+    GenericSeries
     MathematicalExpression
     CoordinateSystemManager
     LocalCoordinateSystem
@@ -139,7 +140,7 @@ import weldx.welding
 
 # class imports to weldx namespace
 from weldx.config import Config
-from weldx.core import MathematicalExpression, TimeSeries
+from weldx.core import GenericSeries, MathematicalExpression, TimeSeries
 from weldx.geometry import (
     ArcSegment,
     Geometry,